Log row message: receive styles from parent
